+++ tests/test_ttlcache.py
@@ -1,17 +1,7 @@
 import unittest
 
-from . import CacheTestMixin
+from . import CacheTestMixin, DecoratorTestMixin
 from cachetools import ExpiredError, TTLCache, ttl_cache
-
-
-@ttl_cache(maxsize=2)
-def cached(n):
-    return n
-
-
-@ttl_cache(maxsize=2, typed=True, lock=None)
-def cached_typed(n):
-    return n
 
 
 class Timer:
@@ -21,17 +11,20 @@ class Timer:
     def __call__(self):
         return self.__time
 
-    def inc(self):
-        self.__time = self.__time + 1
+    def tick(self):
+        self.__time += 1
 
 
-class TTLCacheTest(unittest.TestCase, CacheTestMixin):
+class TTLCacheTest(unittest.TestCase, CacheTestMixin, DecoratorTestMixin):
 
-    def make_cache(self, maxsize, getsizeof=None, ttl=0):
+    def cache(self, maxsize, ttl=0, getsizeof=None):
         return TTLCache(maxsize, ttl, timer=Timer(), getsizeof=getsizeof)
 
-    def test_ttl_insert(self):
-        cache = self.make_cache(maxsize=2, ttl=2)
+    def decorator(self, maxsize, ttl=0, typed=False, lock=None):
+        return ttl_cache(maxsize, ttl, timer=Timer(), typed=typed, lock=lock)
+
+    def test_ttl(self):
+        cache = self.cache(maxsize=2, ttl=2)
         self.assertEqual(cache.ttl, 2)
 
         cache[1] = 1
@@ -39,14 +32,14 @@ class TTLCacheTest(unittest.TestCase, CacheTestMixin):
         self.assertEqual(1, len(cache))
         self.assertEqual(1, cache[1])
 
-        cache.timer.inc()
+        cache.timer.tick()
         cache[2] = 2
 
         self.assertEqual(2, len(cache))
         self.assertEqual(1, cache[1])
         self.assertEqual(2, cache[2])
 
-        cache.timer.inc()
+        cache.timer.tick()
         cache[1]
         cache[3] = 3
 
@@ -55,52 +48,8 @@ class TTLCacheTest(unittest.TestCase, CacheTestMixin):
         self.assertNotIn(2, cache)
         self.assertEqual(3, cache[3])
 
-    def test_ttl_expire(self):
-        cache = self.make_cache(maxsize=3, ttl=0)
-        self.assertEqual(cache.ttl, 0)
-
-        cache[1] = 1
-        self.assertEqual(1, cache[1])
-        cache.timer.inc()
-        with self.assertRaises(ExpiredError):
-            cache[1]
-        cache[2] = 2
-        self.assertEqual(2, cache[2])
-        cache.timer.inc()
-        with self.assertRaises(ExpiredError):
-            cache[2]
-        cache[3] = 3
-        self.assertEqual(3, cache[3])
-
-        cache.expire(1)
-        self.assertNotIn(1, cache)
-        self.assertEqual(3, cache[3])
-
-        cache.expire(2)
-        self.assertNotIn(1, cache)
-        self.assertNotIn(2, cache)
-        self.assertEqual(3, cache[3])
-
-        cache.timer.inc()
-        cache.expire()
-        self.assertEqual(0, len(cache))
-        self.assertNotIn(1, cache)
-        self.assertNotIn(2, cache)
-        self.assertNotIn(3, cache)
-
-    def test_ttl_tuple_key(self):
-        cache = self.make_cache(maxsize=1, ttl=0)
-
-        cache[(1, 2, 3)] = 42
-        self.assertEqual(42, cache[(1, 2, 3)])
-        cache.timer.inc()
-        with self.assertRaises(ExpiredError):
-            cache[(1, 2, 3)]
-        cache.expire()
-        self.assertNotIn((1, 2, 3), cache)
-
-    def test_lru_insert(self):
-        cache = self.make_cache(maxsize=2)
+    def test_lru(self):
+        cache = self.cache(maxsize=2)
 
         cache[1] = 1
         cache[2] = 2
@@ -124,47 +73,46 @@ class TTLCacheTest(unittest.TestCase, CacheTestMixin):
         self.assertEqual(cache[5], 5)
         self.assertNotIn(2, cache)
 
-    def test_lru_getsizeof(self):
-        cache = self.make_cache(maxsize=3, getsizeof=lambda x: x)
+    def test_expire(self):
+        cache = self.cache(maxsize=3, ttl=0)
+        self.assertEqual(cache.ttl, 0)
 
         cache[1] = 1
+        self.assertEqual(1, cache[1])
+        cache.timer.tick()
+        with self.assertRaises(ExpiredError):
+            cache[1]
         cache[2] = 2
-
-        self.assertEqual(len(cache), 2)
-        self.assertEqual(cache[1], 1)
-        self.assertEqual(cache[2], 2)
-
+        self.assertEqual(2, cache[2])
+        cache.timer.tick()
+        with self.assertRaises(ExpiredError):
+            cache[2]
         cache[3] = 3
+        self.assertEqual(3, cache[3])
 
-        self.assertEqual(len(cache), 1)
-        self.assertEqual(cache[3], 3)
+        cache.expire(1)
+        self.assertNotIn(1, cache)
+        self.assertEqual(3, cache[3])
+
+        cache.expire(2)
         self.assertNotIn(1, cache)
         self.assertNotIn(2, cache)
+        self.assertEqual(3, cache[3])
 
-        with self.assertRaises(ValueError):
-            cache[4] = 4
-        self.assertEqual(len(cache), 1)
-        self.assertEqual(cache[3], 3)
+        cache.timer.tick()
+        cache.expire()
+        self.assertEqual(0, len(cache))
+        self.assertNotIn(1, cache)
+        self.assertNotIn(2, cache)
+        self.assertNotIn(3, cache)
 
-    def test_decorator(self):
-        self.assertEqual(cached.cache_info(), (0, 0, 2, 0))
-        self.assertEqual(cached(1), 1)
-        self.assertEqual(cached.cache_info(), (0, 1, 2, 1))
-        self.assertEqual(cached(1), 1)
-        self.assertEqual(cached.cache_info(), (1, 1, 2, 1))
-        self.assertEqual(cached(1.0), 1.0)
-        self.assertEqual(cached.cache_info(), (2, 1, 2, 1))
+    def test_tuple_key(self):
+        cache = self.cache(maxsize=1, ttl=0)
 
-        cached.cache_clear()
-        self.assertEqual(cached(1), 1)
-        self.assertEqual(cached.cache_info(), (2, 2, 2, 1))
-
-    def test_typed_decorator(self):
-        self.assertEqual(cached_typed(1), 1)
-        self.assertEqual(cached_typed.cache_info(), (0, 1, 2, 1))
-        self.assertEqual(cached_typed(1), 1)
-        self.assertEqual(cached_typed.cache_info(), (1, 1, 2, 1))
-        self.assertEqual(cached_typed(1.0), 1.0)
-        self.assertEqual(cached_typed.cache_info(), (1, 2, 2, 2))
-        self.assertEqual(cached_typed(1.0), 1.0)
-        self.assertEqual(cached_typed.cache_info(), (2, 2, 2, 2))
+        cache[(1, 2, 3)] = 42
+        self.assertEqual(42, cache[(1, 2, 3)])
+        cache.timer.tick()
+        with self.assertRaises(ExpiredError):
+            cache[(1, 2, 3)]
+        cache.expire()
+        self.assertNotIn((1, 2, 3), cache)